+++ app/Http/Controllers/API/UserController.php
@@ -3,9 +3,12 @@
 namespace App\Http\Controllers\API;
 
 
+use App\Book;
+use App\Movie;
 use App\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\API\BaseController as BaseController;
+use Illuminate\Support\Facades\Auth;
 use Validator;
 
 
@@ -27,36 +30,21 @@ class UserController extends BaseController
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
-        $input = $request->all();
 
 
-        $validator = Validator::make($input, [
-            'name' => 'required',
-            'email' => 'required'
-        ]);
-
-
-        if($validator->fails()){
-            return $this->sendError('Validation Error.', $validator->errors());
-        }
-
-
-        $user = User::create($input);
-
-
-        return $this->sendResponse($user->toArray(), 'User created successfully.');
+        return $this->sendError( 'This Method is not allowed');
     }
 
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
@@ -65,7 +53,7 @@ class UserController extends BaseController
 
 
         if (is_null($user)) {
-            return $this->sendError('User not found.');
+            return $this->sendError('User not with ' . $id . ' is found.');
         }
 
 
@@ -76,35 +64,30 @@ class UserController extends BaseController
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  \Illuminate\Http\Request $request
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, User $user)
+    public function update(Request $request, $id)
     {
         $input = $request->all();
 
-
-//        $validator = Validator::make($input, [
-//            'name' => 'required',
-//            'email' => 'required'
-//        ]);
-//
-//
-//        if($validator->fails()){
-//            return $this->sendError('Validation Error.', $validator->errors());
-//        }
-
-
-        if(!empty($input['name'])){
-            $user->name = $input['name'];
+        $user = User::find($id);
+        if ($user) {
+            if (Auth::id() == $user->id) {
+                if (!empty($input['name'])) {
+                    $user->name = $input['name'];
+                }
+                if (!empty($input['email'])) {
+                    $user->email = $input['email'];
+                }
+                $user->save();
+            } else {
+                return $this->sendError('Unauthorized action.Your ID is ' . Auth::id() . ". Requested User ID is " . $user->id);
+            }
+        } else {
+            return $this->sendError("User with ID " . $id . " is not found.");
         }
-        if(!empty($input['email'])) {
-            $user->email = $input['email'];
-        }
-        $user->save();
-
-
         return $this->sendResponse($user->toArray(), 'User updated successfully.');
     }
 
@@ -112,15 +95,26 @@ class UserController extends BaseController
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(User $user)
+    public function destroy($id)
     {
-        $user->delete();
+        $user = User::find($id);
+        if ($user) {
+            if (Auth::id() == $user->id) {
+                Book::where('user_id',$id)->delete();
+                Movie::where('user_id',$id)->delete();
+                $user->delete();
+            } else {
+                return $this->sendError('Unauthorized action.Your ID is ' . Auth::id() . ". Requested User ID is " . $user->id);
+            }
+        } else {
+            return $this->sendError("User with ID " . $id . " is not found.");
+        }
 
 
-        return $this->sendResponse($user->toArray(), 'User deleted successfully.');
+        return $this->sendResponse($user->toArray(), 'User and all linked books & movies were deleted successfully.');
     }
 
 
@@ -128,7 +122,7 @@ class UserController extends BaseController
      * showBooks
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function showBooks($id)
@@ -136,12 +130,12 @@ class UserController extends BaseController
         $user = User::find($id);
 
         if ($user != null) {
-            $books=$user->books;
-            if ($books===null || empty($books->toArray())) {
+            $books = $user->books;
+            if ($books === null || empty($books->toArray())) {
                 return $this->sendError('No books were found.');
             }
-        }else{
-            return $this->sendError('No user was found.');
+        } else {
+            return $this->sendError('No user with ID ' . $id . ' was found.');
         }
 
 
@@ -153,7 +147,7 @@ class UserController extends BaseController
      * showMovies
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function showMovies($id)
@@ -161,12 +155,12 @@ class UserController extends BaseController
         $user = User::find($id);
 
         if ($user != null) {
-            $movies=$user->movies;
-            if ($movies===null || empty($movies->toArray())) {
+            $movies = $user->movies;
+            if ($movies === null || empty($movies->toArray())) {
                 return $this->sendError('No movies were found.');
             }
-        }else{
-            return $this->sendError('No user was found.');
+        } else {
+            return $this->sendError('No user with ID ' . $id . ' was found.');
         }
